# perf: optimize ResultSet decoding